+++ lib/svtplay_dl/service/svtplay.py
@@ -9,19 +9,18 @@ from urllib.parse import urljoin, urlparse, parse_qs
 from operator import itemgetter
 
 from svtplay_dl.log import log
-from svtplay_dl.service import Service
+from svtplay_dl.service import Service, MetadataThumbMixin
 from svtplay_dl.utils.text import filenamify
 from svtplay_dl.fetcher.hds import hdsparse
 from svtplay_dl.fetcher.hls import hlsparse
 from svtplay_dl.fetcher.dash import dashparse
 from svtplay_dl.subtitle import subtitle
 from svtplay_dl.error import ServiceError
-from svtplay_dl.utils.http import download_thumbnails
 
 URL_VIDEO_API = "http://api.svt.se/videoplayer-api/video/"
 
 
-class Svtplay(Service):
+class Svtplay(Service, MetadataThumbMixin):
     supported_domains = ['svtplay.se', 'svt.se', 'beta.svtplay.se', 'svtflow.se']
 
     def get(self):
@@ -299,12 +298,3 @@ class Svtplay(Service):
             self.output["episodedescription"] = data["video"]["description"]
         except KeyError:
             pass
-
-    def get_thumbnail(self, options):
-        urls = []
-        if self.output["showthumbnailurl"] is not None:
-            urls.append((True, self.output["showthumbnailurl"]))
-        if self.output["episodethumbnailurl"] is not None:
-            urls.append((False, self.output["episodethumbnailurl"]))
-        if urls:
-            download_thumbnails(self.output, options, urls)
